+++ packages/ui/src/app/pages/Proposals/ProposalPreview.stories.tsx
@@ -46,6 +46,7 @@ type Args = {
   isInDiscussionWhitelist: boolean
   type: ProposalDetailsType
   constitutionality: number
+  exactExecutionBlock: number
   vote1: VoteArg
   vote2: VoteArg
   vote3: VoteArg
@@ -73,6 +74,7 @@ export default {
     isDiscussionOpen: true,
     type: 'SignalProposalDetails',
     constitutionality: 1,
+    exactExecutionBlock: 0,
     vote1: 'None',
     vote2: 'None',
     vote3: 'None',
@@ -156,6 +158,7 @@ export default {
                   status,
                   type: args.type,
                   creator: args.isProposer ? alice : bob,
+                  exactExecutionBlock: args.exactExecutionBlock || null,
 
                   discussionThread: {
                     posts: proposalDiscussionPosts,
@@ -229,6 +232,9 @@ export const FillWorkingGroupLeadOpening: Story = {
 }
 export const FundingRequest: Story = {
   args: { type: 'FundingRequestProposalDetails' },
+}
+export const FundingRequestMultipleRecipients: Story = {
+  args: { type: 'FundingRequestMultipleRecipientsProposalDetails' },
 }
 export const RuntimeUpgrade: Story = {
   args: { type: 'RuntimeUpgradeProposalDetails', constitutionality: 2 },